+++ mypy.ini
@@ -1,0 +1,4 @@
+[mypy]
+
+[mypy-pytest.*]
+ignore_missing_imports = True
